move files and fix phcs errors
